+++ CHANGELOG.md
@@ -1,3 +1,54 @@
+# [5.0.0](https://github.com/iCHEF/gypcrete/compare/v4.4.0...v5.0.0) (2022-05-13)
+
+
+### Bug Fixes
+
+* CI never stop running caused by watch mode ([2da930d](https://github.com/iCHEF/gypcrete/commit/2da930dba2dc161058e3c63bca8fdd1dfb36b6d6))
+* clean code ([d54c000](https://github.com/iCHEF/gypcrete/commit/d54c0009d00b4c2039e11cbe3225d1345998b602))
+* fix code review ([830c305](https://github.com/iCHEF/gypcrete/commit/830c305aeed1d2735b49f11124a98f49b0ee5323))
+* fix props name ([53f3719](https://github.com/iCHEF/gypcrete/commit/53f3719c978a3fae446fa7afd2c6059ae00e1c2b))
+* fix scss file indentation ([176bc6f](https://github.com/iCHEF/gypcrete/commit/176bc6f2f8dfc79521da05fc88b233d77b3c7829))
+* lint error ([e1b1bb1](https://github.com/iCHEF/gypcrete/commit/e1b1bb1e0b01af42ba12064351f711b50aeb95bc))
+* lint error ([1262b03](https://github.com/iCHEF/gypcrete/commit/1262b032f8acf603a2ca1ce7969ee55cd0ab0f1d))
+* lint issue ([488503d](https://github.com/iCHEF/gypcrete/commit/488503d45ce3f5e1c34400bc2598c9d68f6d04a4))
+* linter errors in Radio.scss ([9cf2722](https://github.com/iCHEF/gypcrete/commit/9cf27220d69aa60737c3501823561e9d23c53e70))
+* **package.json:** fix generate-icon-components script ([7b95c83](https://github.com/iCHEF/gypcrete/commit/7b95c8343ae7309ae1037029e19be46c993c6099))
+* remove build job ([94dbe0d](https://github.com/iCHEF/gypcrete/commit/94dbe0d64695e457b86f76c465fa73ba6f057c8c))
+* scss stylelint errors ([ff68819](https://github.com/iCHEF/gypcrete/commit/ff688194a0f201660cfc8804adda03474be1e4c9))
+* should not asideControlClickableOnDisabled if not disabled ([673ad51](https://github.com/iCHEF/gypcrete/commit/673ad5168617d7f16bd58de6ce136dca1550a588))
+
+
+### Features
+
+* add .nvmrc for node version ([4c0a1b9](https://github.com/iCHEF/gypcrete/commit/4c0a1b9c28dc1249fec1c5a5fd5aa27df9bf95c7))
+* add "conventional-changelog-cli to package.json ([ad22c1f](https://github.com/iCHEF/gypcrete/commit/ad22c1f6d5d7c3138c6bbd8e2a2ca6e74b53585c))
+* add `add-location` icon ([39aa011](https://github.com/iCHEF/gypcrete/commit/39aa011f88d55aa26b5a9bfd79ebfdc6f5e1d23a))
+* add canary release workflow ([fa17c3d](https://github.com/iCHEF/gypcrete/commit/fa17c3db8a413e3d73b429e7cf5159fc2246d72f))
+* add ci workflow with codecov ([5f56f15](https://github.com/iCHEF/gypcrete/commit/5f56f15384e0ec0d01d96bb55aa71fa4b80eed7d))
+* add Location icon ([d7a5cc3](https://github.com/iCHEF/gypcrete/commit/d7a5cc33f6be93c293ca0c0fe7eb324d7caadea9))
+* add props for RadioSelectList ([984dedc](https://github.com/iCHEF/gypcrete/commit/984dedccdd7a0655de582ccde184000b5c94fd02))
+* add radio-empty.svg and radio-selected.svg ([ad75ef6](https://github.com/iCHEF/gypcrete/commit/ad75ef6e0a11195899a657a5e9d35bba39ce827c))
+* add release workflow ([6620403](https://github.com/iCHEF/gypcrete/commit/66204034d4a1ea59a3d73d26721c609b6b4362aa))
+* add SvgRadioEmpty and SvgRadioSelected ([cdfb6d4](https://github.com/iCHEF/gypcrete/commit/cdfb6d4f4b269748675788f40f7780d2710f6911))
+* change `fill` value to currentColor ([b2e51ca](https://github.com/iCHEF/gypcrete/commit/b2e51ca1bbe82801a1ee510ee9a746e97f810e23))
+* change AddLocation icon ([8ad1d7b](https://github.com/iCHEF/gypcrete/commit/8ad1d7b18587622ce8ccedf1c4f11a697c882c34))
+* change Location icon ([ad2600b](https://github.com/iCHEF/gypcrete/commit/ad2600b299f7f482b3ef96a951145369e6a01a79))
+* **CheckboxEmpty:** rename radio-empty with checkbox-empty ([369a6b3](https://github.com/iCHEF/gypcrete/commit/369a6b387f9fcee626221813a16acf83277f741e))
+* **CheckboxHalf:** rename to CheckboxHalf ([33cf1da](https://github.com/iCHEF/gypcrete/commit/33cf1da858fd6a78a76c22714012c011ecfafe0a))
+* **CheckboxSelected:** rename to CheckboxSelected ([0883c8d](https://github.com/iCHEF/gypcrete/commit/0883c8da39c77caee12ec2aa6857e4a725f6f5d6))
+* deleted unused deploy.sh ([c06773e](https://github.com/iCHEF/gypcrete/commit/c06773e407832ecddc50cdb12ff434b7ee0c051b))
+* deleted unused travis cd ([c4a99b1](https://github.com/iCHEF/gypcrete/commit/c4a99b1ab8e39c9f6a5dd1d9149e488f1a03e03f))
+* export RadioSelectList and RadioSelectOption ([825d0b1](https://github.com/iCHEF/gypcrete/commit/825d0b1f8ad826ebb972929b62a55524e1b365c0))
+* implement Radio component ([df4e347](https://github.com/iCHEF/gypcrete/commit/df4e347d603802fe1fb8ab6132e5928b02619424))
+* implement RadioSelectList and RadioSelectOption ([d595eaf](https://github.com/iCHEF/gypcrete/commit/d595eafc16ad41a435a5970fdd6df1956ff8eb71))
+* remove fixed `fill` color ([87ba9ce](https://github.com/iCHEF/gypcrete/commit/87ba9cedef64c83897925cf9f7901ffe4b336344))
+* specify node maximum version in package.json ([86aee36](https://github.com/iCHEF/gypcrete/commit/86aee360d2b1daf171722a93c7277129b25db4bb))
+* update release workflow ([00fba6d](https://github.com/iCHEF/gypcrete/commit/00fba6d9741fb8b6fb90e38ba90f924a3481a69e))
+* update release workflow with `iCHEF/npm-install-with-cache-action@v1` ([40cfe3a](https://github.com/iCHEF/gypcrete/commit/40cfe3ad3f04b2eb00215235f9447c1838be297d))
+* use ichef npm-install-with-cache-action to install packages ([3ff5ed1](https://github.com/iCHEF/gypcrete/commit/3ff5ed136f5f5c73fde14cbb545b5be5725c9bfa))
+
+
+
 # Change Log
 All notable changes to this project will be documented in this file.